Separated Classes
Hopefully this compiles in Clang.

Changelog:
- Moved CustomSorts class into customsorts.h
- Renamed customimplementations.h & .cpp to minheap.h & .cpp
- Commented out sorting function in PrimeWin
