+++ Baseball-Itinerary-Program/minheap.h
@@ -1,50 +1,12 @@
-#ifndef CUSTOMIMPLEMENTATIONS_H
-#define CUSTOMIMPLEMENTATIONS_H
+#ifndef MINHEAP_H
+#define MINHEAP_H
 #include <deque>
-#include <stdexcept>
 #include <string>
-#include <sstream>
-#include <vector>
-#include <cstddef>
-#include <QString>
 #include <utility>
-#include <QDebug>
+#include <stdexcept>
+#include <sstream>
 using namespace std;
 
-/*************************************************************************
- * CustomSorts
- * ----------------------------------------------------------------------
- * This class stores all methods implemented by developers used for
- * sorting data.
- * The SortObj class is a helper class that will associate a stadNum
- * with whatever data that needs to be sorted (for now it's just QString
- * but maybe someone could make it a template *hint *HINT) ¯\_(ツ)_/¯
- * ----------------------------------------------------------------------
- * PRE-REQUISITES
- * ----------------------------------------------------------------------
- * #include <vector>
- * using namespace std;
- ************************************************************************/
-class SortObj   //Objects to sort that also store an associated stadNum
-{
-public:
-    int stadNum;
-    QString sortee;
-
-    //Dereference to get the value to be sorted
-    QString operator *() const {return sortee;}
-    //Pointer operator to get stadNum
-    int operator &() const {return stadNum;}
-};
-class CustomSorts
-{
-public:
-    CustomSorts();
-    ~CustomSorts();
-
-    //Insertion sort
-    vector<SortObj> InsertionSort(vector<SortObj> sortThese);
-};
 /*************************************************************************
  * MinMeap (Minimum Binary Heap with Map Parallel) (Int Specific)
  * ----------------------------------------------------------------------
@@ -95,8 +57,4 @@ private:
     int lChildOf(int index);             //Get left child index
     int rChildOf(int index);             //Get right child index
 };
-
-
-
-
-#endif // CUSTOMIMPLEMENTATIONS_H
+#endif // MINHEAP_H
